Split out datasource-related stuff in pydantic_models
