+++ s7/pydantic_models/soft7_entity.py
@@ -5,16 +5,13 @@ from __future__ import annotations
 import json
 import logging
 import sys
-import traceback
 from pathlib import Path
 from typing import (
     TYPE_CHECKING,
     Annotated,
     Any,
     Optional,
-    Protocol,
     Union,
-    runtime_checkable,
 )
 
 if sys.version_info >= (3, 10):
@@ -28,12 +25,9 @@ from pydantic import (
     AliasChoices,
     AnyUrl,
     BaseModel,
-    ConfigDict,
     Field,
-    TypeAdapter,
     ValidationError,
 )
-from pydantic.functional_serializers import model_serializer
 from pydantic.functional_validators import (
     field_validator,
     model_validator,
@@ -44,9 +38,6 @@ from pydantic_core import Url
 from s7.exceptions import EntityNotFound
 
 if TYPE_CHECKING:  # pragma: no cover
-    from pydantic import SerializerFunctionWrapHandler
-    from pydantic.fields import FieldInfo
-
     UnshapedPropertyType = Union[
         str, float, int, complex, dict, bool, bytes, bytearray, BaseModel
     ]
@@ -87,23 +78,6 @@ def SOFT7IdentityURI(url: str) -> SOFT7IdentityURIType:
         return SOFT7IdentityURIType(url)
 
     raise TypeError(f"Expected str or Url, got {type(url)}")
-
-
-@runtime_checkable
-class GetData(Protocol):
-    """Get a named datum (property or dimension) from the data resource.
-
-    Properties:
-        soft7_property: The name of a datum to get, i.e., the SOFT7 data resource
-            property (or dimension).
-
-    Returns:
-        The value of the SOFT7 data resource property (or dimension).
-
-    """
-
-    def __call__(self, soft7_property: str) -> Any:  # pragma: no cover
-        ...
 
 
 SOFT7EntityPropertyType = Union[
@@ -194,163 +168,6 @@ def parse_identity(identity: AnyUrl) -> tuple[AnyUrl, Optional[str], str]:
     namespace += identity.path.rstrip("/")[: -len(version) - len(name) - 2]
 
     return AnyUrl(namespace), version or None, name
-
-
-class CallableAttributesBaseModel(BaseModel):
-    """Call, resolve, and cache attributes if they are SOFT7 properties."""
-
-    _resolved_fields: dict[str, PropertyType] = {}
-
-    def __getattribute__(self, name: str) -> Any:
-        """Get an attribute.
-
-        This function will _always_ be called whenever an attribute is accessed.
-
-        Having this function allows us to intercept the attribute access and if the
-        attribute is a SOFT7 property, we can retrieve the value from the underlying
-        data source, i.e., call the `__get_data()` function that the attribute value
-        currently is.
-
-        Before calling the `__get_data()` function, we need to ensure the attribute
-        is not a SOFT7 metadata attribute, i.e., starts with `soft7___`.
-
-        """
-        resolved_attr_value = "NOT SET"
-
-        try:
-            if (
-                # If dunder method, return as-is.
-                # or is a SOFT7 metadata field
-                name.startswith(("__", "soft7___"))
-                # Or otherwise a non-SOFT7 attribute
-                or name
-                not in (model_fields := object.__getattribute__(self, "model_fields"))
-                # Or if private attributes are not set (not yet properly initialized)
-                or not (
-                    private_attrs := object.__getattribute__(
-                        self, "__pydantic_private__"
-                    )
-                )
-            ):
-                return object.__getattribute__(self, name)
-
-            # Check "cache"
-            # Retrieve from "cache"
-            if name in (resolved_fields := private_attrs.get("_resolved_fields", {})):
-                LOGGER.debug("Using cached value for %s", name)
-                # If the field has already been resolved, use the resolved value.
-                return resolved_fields[name]
-
-            # Retrieve from data source
-            LOGGER.debug("Resolving value for %s", name)
-            resolved_attr_value = object.__getattribute__(self, name)(
-                soft7_property=name
-            )
-
-            # Use TypeAdapter to return and validate the value against the
-            # generated type. This effectively validates the shape and
-            # dimensionality of the value, as well as the inner most expected type.
-            field_info: FieldInfo = model_fields[name]
-            resolved_fields[name] = TypeAdapter(
-                field_info.rebuild_annotation()
-            ).validate_python(resolved_attr_value)
-
-            BaseModel.__setattr__(self, "_resolved_fields", resolved_fields)
-
-            return resolved_fields[name]
-
-        except ValidationError as exc:
-            LOGGER.error(
-                "Attribute: %s. Data: %r\n%s: %s\nTraceback:\n%s",
-                name,
-                resolved_attr_value,
-                exc.__class__.__name__,
-                exc,
-                "".join(traceback.format_tb(exc.__traceback__)),
-            )
-            raise AttributeError(
-                f"Could not type and shape validate attribute {name!r} from the data "
-                "source."
-            ) from exc
-
-        except Exception as exc:
-            LOGGER.error(
-                "An error occurred during attribute resolution:\n%s: %s\n"
-                "Traceback:\n%s",
-                exc.__class__.__name__,
-                exc,
-                "".join(traceback.format_tb(exc.__traceback__)),
-            )
-            raise AttributeError(f"Could not retrieve attribute {name!r}.") from exc
-
-    @model_serializer(mode="wrap", when_used="always")
-    def _serialize_callable_attributes(
-        self, handler: SerializerFunctionWrapHandler
-    ) -> Any:
-        """Serialize all "lazy" SOFT7 property values.
-
-        If the value matches the GetData protocol, i.e., it's a callable function with
-        the `name` parameter, call it with the property's name and the result will be
-        used in a copy of the model. Otherwise, the value will be used as-is.
-
-        The copy of the model is returned through the SerializerFunctionWrapHandler.
-        """
-        if not isinstance(self, BaseModel):
-            raise TypeError(
-                "This mixin class may only be used with pydantic.BaseModel subclasses."
-            )
-
-        # This iteration works, due to how BaseModel yields fields (from __dict__ +
-        # __pydantic__extras__).
-        for field_name, field_value in self:
-            if field_name in self._resolved_fields:
-                # If the field has already been resolved, use the resolved value.
-                continue
-
-            if isinstance(field_value, GetData):
-                # Call the function via self.__getattribute__ to ensure proper type
-                # validation and store the returned value for later use.
-                self._resolved_fields[field_name] = getattr(self, field_name)
-
-            # Else: The value is not a GetData, so use it as-is, i.e., no changes.
-
-        return handler(self.model_copy(update=self._resolved_fields))
-
-
-class SOFT7DataSource(CallableAttributesBaseModel):
-    """Generic SOFT7 data source
-
-    This doc-string should be replaced with the specific data source's `description`.
-
-    The configuration options:
-    - `extra="forbid"`: Ensures an exception is raised if the instantiated data source
-      tries to specify undefined properties.
-    - `frozen=True`: Ensures an exception is raised if the instantiated data source
-      tries to modify any properties, i.e., manually set an attribute value.
-    - `validate_default=False`: Set explicitly (`False` is the default) to avoid a
-      ValidationError when instantiating the data source. This is due to the properties
-      being lazily retrieved from the data source.
-
-    Note:
-        The defined `soft7___*` fields will be overwritten by `pydantic.create_model()`,
-        but they are defined here to help with understanding what the SOFT7 metadata
-        attributes are.
-
-        They act as a "contract" between the factory function and the SOFT7DataSource
-        model, i.e., the factory function must return a dictionary that at least defines
-        the metadata attributes, since otherwise a ValidationError will be raised upon
-        instantiation.
-
-    """
-
-    model_config = ConfigDict(extra="forbid", frozen=True, validate_default=False)
-
-    soft7___dimensions: BaseModel
-    soft7___identity: AnyUrl
-
-    soft7___namespace: str
-    soft7___version: Optional[str]
-    soft7___name: str
 
 
 class SOFT7EntityProperty(BaseModel):
